+++ internal/cli/service_uid_unix.go
@@ -1,0 +1,7 @@
+//go:build !windows
+
+package cli
+
+import "os"
+
+func currentUID() int { return os.Getuid() }